Run Envision only once in version control evaluation scripts
A single instance is used for multiple evalutions. The presence of control
files is used to signal that new input should be processed.

+++ misc/version-control/scripts/diff_git_dev.sh
@@ -5,7 +5,7 @@
 # If the versions are different, an entry specifying the revision and file is written to merges/issues_git.
 # Additionally, it writes an entry for all encountered files to merges/all which is helpful when using the comm command with these kind of files.
 
-rm "${1}/merges/issues_git"
+rm -rf "${1}/merges/issues_git"
 
 merges="${1}/merges/*"
 for m in $merges; do
@@ -15,7 +15,7 @@ for m in $merges; do
 		if [ -d "${fdir}" ]; then
 			(
 				cd $fdir
-				if [ -f base.java ] && [ -f dev.java ] && [ -f master.java ]; then
+				if [ -f base.java ] && [ -f dev.java ] && [ -f master.java ] && [ -f devMerged.java ]; then
 					cp master.java gitMerged.java
 					git merge-file -L master.java --quiet gitMerged.java base.java dev.java
 					diff devMerged.java gitMerged.java > diff_dev_git
